migration: Fix rate limit
The migration thread appears to want to allow writeout to occur at full
speed rather than being rate limited during completion of state saving,
but sets the limit to INT_MAX when xfer_limit is INT64_MAX. This causes
problems if there's more than 2GB of state left to save at this point. It
probably ought to just be INT64_MAX instead.

(cherry picked from commit 40596834c0d57a223124a956ccbe39dfeadc9f0e)


Conflicts:
	migration.c

+++ migration.c
@@ -687,7 +687,7 @@ static void *migration_thread(void *opaque)
                 ret = vm_stop_force_state(RUN_STATE_FINISH_MIGRATE);
                 if (ret >= 0) {
                     if (!migrate_postcopy_outgoing()) {
-                        qemu_file_set_rate_limit(s->file, INT_MAX);
+                        qemu_file_set_rate_limit(s->file, INT64_MAX);
                     }
                     qemu_savevm_state_complete(s->file, &s->params);
                 }
